Change semantics and timing of failure/success patterns
Matching of those patterns does not include the characters successfully
matching normal patterns. When a normal pattern is matched only the
characters before the match is matched for failure/success.

This means for example that it now is possible to implement a test case
which explicit matches the prompt and fails if there are more prompts
than those that are explicitly matched.

This test case will detect the spurious prompt after "bar":

  -SH-PROMPT:
  !foo
  ?some output
  ?SH-PROMPT:
  !bar
  ?some output
  !baz
  ?some output
  ?SH-PROMPT: